desktop/requirements.txt
PySide6==6.7.0
aiohttp==3.9.5
PyYAML==6.0.1
